Empty video folder handling
Handling of empty folders data/new and data/verified. Right now, this happens when recording a video, which looks like a seperate bug.
New handling does not display empty folders as videos in the program GUI.

Also added new test under tests/test.state.py for this commits empty video folder handling.

Addtionally, now enforcing type int to key k in self.detections in instance of State. This prevents crash when drawing rectangles on image due to comparison of strs and ints.

Lastly, no .pre-commit for this one, as there are errors relating to the script for my machine. Will be fixed for future pushes